LPD-53232 Refactor module name to reflect accurate COMMERCE ownership
Widget Configuration: Permissions belongs to COMMERCE & USM
