+++ modules/test/playwright/tests/portlet-configuration-permissions-web/pagePermissions.spec.ts
