+++ modules/test/playwright/tests/portlet-configuration-permissions-web/rolePermissions.spec.ts
