+++ modules/test/playwright/tests/portlet-configuration-permissions-web/test.properties
